[FIX] Fix issues in test applications

+++ test/tests/test01.c
@@ -63,7 +63,7 @@ ujoBool test01()
 	err = ujo_writer_add_int64(ujow, 12345678);
 	print_return_ujo_err(err,"ujo_writer_add_int64");
 
-	err = ujo_writer_add_float16(ujow, 43332);
+	err = ujo_writer_add_float16(ujow, (float32_t)43332.0);
 	print_return_ujo_err(err,"ujo_writer_add_float16");
 
 	err = ujo_writer_add_float32(ujow, (float32_t)3.14);
