+++ test/tests/test08.c
@@ -56,13 +56,13 @@ ujoBool test08()
 	err = ujo_writer_list_open(ujow);
 	print_return_ujo_err(err,"ujo_writer_list_open"); 
 
-	err = ujo_writer_add_float16(ujow, 4.13);
+	err = ujo_writer_add_float16(ujow, (float32_t)4.13);
 	print_return_ujo_err(err,"ujo_writer_add_float16");
 
-	err = ujo_writer_add_float32(ujow, 3.14);
+	err = ujo_writer_add_float32(ujow, (float32_t)3.14);
 	print_return_ujo_err(err,"ujo_writer_add_float32"); 
 
-	err = ujo_writer_add_float64(ujow, 3.14);
+	err = ujo_writer_add_float64(ujow, (float64_t)3.14);
 	print_return_ujo_err(err,"ujo_writer_add_float64"); 
 
 	err = ujo_writer_list_close(ujow);
